requirements.txt
# Core data science libraries
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
wordcloud>=1.8.0

# NLP and ML
transformers>=4.20.0
torch>=1.12.0
datasets>=2.0.0
evaluate>=0.3.0
nltk>=3.7
spacy>=3.4.0
gensim>=4.2.0
vaderSentiment>=3.3.2

# Statistics
statsmodels>=0.13.0
scipy>=1.9.0

# API and web requests
openai>=0.27.0
requests>=2.28.0
tenacity>=8.2.0

# Environment and utilities
python-dotenv>=0.19.0
pathlib2>=2.3.0

# Optional: for better performance
# accelerate>=0.12.0
# bitsandbytes>=0.35.0 